move tests to their client tests
