+++ internal/extsvc/github/v3_test.go
@@ -15,10 +15,22 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"github.com/sourcegraph/sourcegraph/internal/extsvc/auth"
+	"github.com/sourcegraph/sourcegraph/internal/httpcli"
 	"github.com/sourcegraph/sourcegraph/internal/httptestutil"
 	"github.com/sourcegraph/sourcegraph/internal/rcache"
 	"github.com/sourcegraph/sourcegraph/internal/testutil"
 )
+
+func newTestClient(t *testing.T, cli httpcli.Doer) *V3Client {
+	return newTestClientWithAuthenticator(t, nil, cli)
+}
+
+func newTestClientWithAuthenticator(t *testing.T, auth auth.Authenticator, cli httpcli.Doer) *V3Client {
+	rcache.SetupForTest(t)
+
+	apiURL := &url.URL{Scheme: "https", Host: "example.com", Path: "/"}
+	return NewV3Client("Test", apiURL, auth, cli)
+}
 
 func TestNewRepoCache(t *testing.T) {
 	cmpOpts := cmp.AllowUnexported(rcache.Cache{})
@@ -476,6 +488,20 @@ func TestGetRepository(t *testing.T) {
 			t.Fatal("expected repo, but got nil")
 		}
 	})
+
+	t.Run("repo not found", func(t *testing.T) {
+		repo, err := cli.GetRepository(context.Background(), "owner", "repo")
+		if !IsNotFound(err) {
+			t.Errorf("got err == %v, want IsNotFound(err) == true", err)
+		}
+		if err != ErrRepoNotFound {
+			t.Errorf("got err == %q, want ErrNotFound", err)
+		}
+		if repo != nil {
+			t.Error("repo != nil")
+		}
+	})
+
 }
 
 // ListOrganizations is primarily used for GitHub Enterprise clients. As a result we test against
@@ -754,3 +780,144 @@ func newV3TestEnterpriseClient(t testing.TB, name string) (*V3Client, func()) {
 }
 
 func strPtr(s string) *string { return &s }
+
+func TestClient_ListRepositoriesForSearch(t *testing.T) {
+	mock := mockHTTPResponseBody{
+		responseBody: `
+{
+  "total_count": 2,
+  "incomplete_results": false,
+  "items": [
+    {
+      "node_id": "i",
+      "full_name": "o/r",
+      "description": "d",
+      "html_url": "https://github.example.com/o/r",
+      "fork": true
+    },
+    {
+      "node_id": "j",
+      "full_name": "a/b",
+      "description": "c",
+      "html_url": "https://github.example.com/a/b",
+      "fork": false
+    }
+  ]
+}
+`,
+	}
+	c := newTestClient(t, &mock)
+
+	wantRepos := []*Repository{
+		{
+			ID:            "i",
+			NameWithOwner: "o/r",
+			Description:   "d",
+			URL:           "https://github.example.com/o/r",
+			IsFork:        true,
+		},
+		{
+			ID:            "j",
+			NameWithOwner: "a/b",
+			Description:   "c",
+			URL:           "https://github.example.com/a/b",
+			IsFork:        false,
+		},
+	}
+
+	reposPage, err := c.ListRepositoriesForSearch(context.Background(), "org:sourcegraph", 1)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if !repoListsAreEqual(reposPage.Repos, wantRepos) {
+		t.Errorf("got repositories:\n%s\nwant:\n%s", stringForRepoList(reposPage.Repos), stringForRepoList(wantRepos))
+	}
+}
+
+func TestClient_ListRepositoriesForSearch_incomplete(t *testing.T) {
+	mock := mockHTTPResponseBody{
+		responseBody: `
+{
+  "total_count": 2,
+  "incomplete_results": true,
+  "items": [
+    {
+      "node_id": "i",
+      "full_name": "o/r",
+      "description": "d",
+      "html_url": "https://github.example.com/o/r",
+      "fork": true
+    },
+    {
+      "node_id": "j",
+      "full_name": "a/b",
+      "description": "c",
+      "html_url": "https://github.example.com/a/b",
+      "fork": false
+    }
+  ]
+}
+`,
+	}
+	c := newTestClient(t, &mock)
+
+	// If we have incomplete results we want to fail. Our syncer requires all
+	// repositories to be returned, otherwise it will delete the missing
+	// repositories.
+	_, err := c.ListRepositoriesForSearch(context.Background(), "org:sourcegraph", 1)
+
+	if have, want := err, ErrIncompleteResults; want != have {
+		t.Errorf("\nhave: %s\nwant: %s", have, want)
+	}
+}
+
+func TestClient_ListOrgRepositories(t *testing.T) {
+	mock := mockHTTPResponseBody{
+		responseBody: `[
+  {
+    "node_id": "i",
+    "full_name": "o/r",
+    "description": "d",
+    "html_url": "https://github.example.com/o/r",
+    "fork": true
+  },
+  {
+    "node_id": "j",
+    "full_name": "o/b",
+    "description": "c",
+    "html_url": "https://github.example.com/o/b",
+    "fork": false
+  }
+]
+`,
+	}
+
+	c := newTestClient(t, &mock)
+	wantRepos := []*Repository{
+		{
+			ID:            "i",
+			NameWithOwner: "o/r",
+			Description:   "d",
+			URL:           "https://github.example.com/o/r",
+			IsFork:        true,
+		},
+		{
+			ID:            "j",
+			NameWithOwner: "o/b",
+			Description:   "c",
+			URL:           "https://github.example.com/o/b",
+			IsFork:        false,
+		},
+	}
+
+	repos, hasNextPage, _, err := c.ListOrgRepositories(context.Background(), "o", 1, "all")
+	if err != nil {
+		t.Fatal(err)
+	}
+	if !repoListsAreEqual(repos, wantRepos) {
+		t.Errorf("got repositories:\n%s\nwant:\n%s", stringForRepoList(repos), stringForRepoList(wantRepos))
+	}
+	if !hasNextPage {
+		t.Errorf("got hasNextPage: false want: true")
+	}
+}
